cmd/clef, signer: make fourbytes its own package, break dep cycle
